+++ admin/tabs/clients/activity/ClientDetailModel.ts
@@ -1,10 +1,10 @@
-import {ClientsModel} from '@xh/hoist/admin/tabs/client/clients/ClientsModel';
+import {ClientsModel} from '../ClientsModel';
 import {ColumnSpec} from '@xh/hoist/cmp/grid';
 import {HoistModel, LoadSpec, lookup, PlainObject, XH} from '@xh/hoist/core';
 import {StoreRecord} from '@xh/hoist/data';
 import {bindable, computed, makeObservable} from '@xh/hoist/mobx';
 import {ReactNode} from 'react';
-import {ActivityDetailProvider} from '../../../activity/tracking/detail/ActivityDetailModel';
+import {ActivityDetailProvider} from '../../activity/tracking/detail/ActivityDetailModel';
 
 export class ClientDetailModel extends HoistModel implements ActivityDetailProvider {
     @lookup(ClientsModel) clientsModel: ClientsModel;